channel signals
